Change attribute order for examples
Use the following order:

1. class
2. id
3. data-*
4. Everything else

HTML attributes should be listed in an order that reflects the fact
that class names are the primary interface through which CSS and
JavaScript select elements.

https://github.com/necolas/idiomatic-html#attribute-order

This also makes it clear in the examples which classes are be used to
apply the GOV.UK elements styles for each example.