Improve date pattern
(Following discussion on the
Hackpad)[https://designpatterns.hackpad.com/Dates-vpx6XlVjIbE]  and a
conversation with @matharden about user research from Visit someone in
prison, keep three inputs for date. Use the pattern and number
attributes to trigger a numeric keyboard and hide the “stepper” using
css.

This removes the need for Modernizr, so this script is deleted.

Also reformat the _form_date.scss file to meet the GOV.UK style guide
recommendations for Sass nesting:
https://github.com/alphagov/styleguides/blob/master/css.md