Improve details summary polyfill
This wasn’t correctly detecting the “open” attribute.

This is a quick fix to get feedback from Leonie and *needs* refactoring
and tests.

For now, it has two checks - first for native support, which uses the
boolean ‘open’ attribute.  Then for non-native support, which checks
for an ‘open’ attribute with an empty value, otherwise ‘open’ is
undefined (not present) and the details content should be closed.

This check is also used to either create an open or closed arrow
(pointing right or down) depending on the initial ‘open’ state.