Remove duplicate errors, change error position
Having more than one error message per question is more likely to lead
to mistakes in error text (if they are not both kept updated), both
messages being read out, or neither.

Instead use a single error message, placed between the label or legend
text and the form input.

Also show the summary box for all examples. This is the fastest and
best way to find and correct errors and should be consistent regardless
of how many questions appear on a page.