Toggle the open attribute when opened or closed
To match the behaviour of the native details element, for non-native
details, the open attribute must be removed when the widget is closed
and set to ‘open’ when the widget is opened.

http://www.w3.org/html/wg/drafts/html/master/semantics.html#the-details-
element

The user agent should allow the user to request that the additional
information be shown or hidden. To honor a request for the details to
be shown, the user agent must set the open attribute on the element to
the value open. To honor a request for the information to be hidden,
the user agent must remove the open attribute from the element.