Make inputs work without extra code
In order for the inputs to function in code, they must have the `name` attribute. I've added this so that people copying the snippets don't have to add any extra code.

For inputs in a series (checkboxes) I've used the `variable[]` naming convention from PHP, but I don't know if that works in Node/Express.