Add missing form elements, fieldsets and legends
For some of the snippets, these were only showing the form elements for
that example.

So these make sense, add context for each example. Following “One thing
per page” each example should have a heading, a fieldset and a legend -
where it is grouping a set of related form controls.