Create a generic 'notice' wrapper
This is a clearing wrapper, relatively positioned, so an icon can sit
next to the text and be vertically aligned.

At present this will only work with icons which are 35px x 35px (the
size of the important icon). This icon should be resized to be 36px.

The class to set the bold text inside the notice wrapper is set on the
paragraph itself, so this ‘notice’ wrapper could also be used to align
text in different sizes with an icon on the left.