Remove Firefox "fix" for arrow edges
This was an attempt to smooth the edges of the left pointing triangle
in Firefox, as they appeared jagged.

Unfortunately, this caused the following error in the govuk Rails
prototyping kit: Base-level rules cannot contain the
parent-selector-referencing character '&'.

Since this “fix” is no longer required, removing these rules fixes this.