Don't bump the font-size on the main element to 19px
Don’t set the base font-size to 19px for the main content area.

A later commit will add parent classes to wrap common elements,
to ensure the size for body copy is 19px.

The govuk_template sets the base font size to be 16px, update the em
function in helpers to use 16px its context.

Where the em function is used and the second argument is 16, remove it
as this is now the default.

Where the em function is used and the context was previously 19px,
adjust the sizing accordingly.