Scope all paragraphs to the parent class .text
Paragraph elements are the only element selector used for styling in
the typography partial.

Scope these to the parent selector .text, which will be used to wrap
blocks of text.

Also move the max-width setting to paragraphs within these .text blocks.

This will also enable anyone using markdown, or not wishing to add
class attributes to each element in a text block to use the same
pattern, for example, for headings and lists.

An example of this is shown in elements-page.scss where list items are
scoped to .text to set a max-width, as govuk elements doesn’t use a
2/3, 1/3 grid (because of its full-width examples) so elements within
.text must have a max-width set.

The only other way to achieve this would be to use a 2/3 grid column
for text blocks within govuk elements and full width column for each
example.