Scope all paragraphs to the parent class .text
Bump up the font size to 19px on .text, otherwise this needs redefining
all over the place.

This will require a wrapper with the class of .text to be used to set a
19px default for any elements within .text.

Move the max-width setting to paragraphs within these .text blocks.

This will also enable anyone using markdown, or not wishing to add
class attributes to each element in a text block to use the same
pattern, for example, for headings and lists.

The elements page stylesheet also sets a max-width for lists, to avoid
having to use grid layout to constrain the text on its pages.

This will also allow the prototype kit to have an additional stylesheet
which can extend common element styles, using the .text scope.