Remove the default value for list style type
This is a fix to remove the defaults (disc, or decimals) from unordered
and ordered lists. These will require the .list-bullet and .list-number
classes, to set spacing and apply a list-style.

This removes the override to remove these within the elements example
boxes.

It also resets the bottom margin for bulleted lists within elements
example boxes to 20px - as this is overridden by elements example page
stylesheet to give bulleted lists a larger bottom margin.