Whitelist only the Sass folder
The files within public/images are used exclusively by this
application, so will not need to be shared - aside from the breadcrumb
separator image, which will be moved to the front end toolkit’s images
folder.

The JavaScript does not follow the govuk module pattern, any modules to
be shared should sit in the govuk frontend toolkit.