Fix Sass linting
Ensure there is a single line per selector
Add a space after a comma
Add a space after a colon

This error remains for this file:  _base.scss line 8 - NameFormat: Name
of variable `NTA-Light` should be written in all lowercase letters with
hyphens instead of underscores.

This file should also be renamed to _base-govuk-template.scss as the
contents of this file have been copied from the govuk template.