Add consistent bottom margins
For body text and lists, panels and tables:
15px for smaller screens, 20px for larger screens - to match the
spacing under body text.

For large headings - make the margins consistent with the govspeak
titles - for both 48px and 36px titles, set a top margin of 15px for
smaller viewports and 45px for larger viewports.

For smaller headings, use 30px as a bottom margin and for the headings
which are the same size as body copy, but bold - use the same spacing
for paragraphs and lists.