Centre text on full width buttons
Text on native input button/submit inputs is normally centred. When links are dressed as buttons the text is left aligned. This is okay on tablet views and above as buttons are only as wide as they need to be but on mobile views, buttons are full width. Centre the text for full width buttons for consistency.