Add a deprecation warning for external links
We’d potentially like to remove external link styling, but first we want to see if anyone else has any input into whether they are useful or not. This adds a deprecation warning. I’ll also add a note to the changelog when we release the next version, and publicise it at that point.