Make sure hover state doesn’t ‘stick’ after deselection
iOS will fake :hover states after touching to make sure that pure-CSS :hover menus still work. This has the effect on the custom checkboxes that a grey checkmark is left in place after a checkbox is deselected.

Luckily, CSS4 media queries has a hover ability query that Safari and Chrome support (Safari 9.1+, Chrome 41+). We can use this to override the grey checkmark state and make sure it’s hidden after a touch deselection.