Remove wrapping form elements
These were added to each of the examples as the selection-buttons.js
requires a parent form. Since then, the snippets have been modified so
they now show how they would work with a “one question per page” form
and also include a heading, then the parent form with a fieldset and
hidden legend for assistive technology.

The form elements wrapping these examples are no longer required, so
can be removed.