Inherit font stack from toolkit
This commit removes the hard-coded reference to New Transport in
the GOV.UK template base, and replaces it with a mostly equivalent
variable used within the Frontend Toolkit, which can be overridden.

Because the Frontend Toolkit is a dependency of GOV.UK elements, the
new variable is usually initialised to MTA Light anyway, but in a way
that eases the process of reusing the template if one is unable to use
the New Transport font.