Fix radios in Windows high contrast mode
Windows has a high contrast theme that sets borders and background to different colours. In our case we’ve rendered the custom checkboxes using only borders, so that’s fine. The custom radio however has its outline rendered with a border, but the bullet rendered with a background. In high contrast mode the bullet was black-on-black and wasn’t showing.

We can fix this by rendering the bullet with borders instead. We just set a width and height of 0, then make the border 10px thick on every side.

Tested in IE8+, Fx, Chrome, old iOS Safari and IE Mobile.