Get rid of potential 300ms pause with radios in mobile browsers
Modern mobile browsers can disable the pause programmatically with touch-action: manipulation (thanks to @patrickhlauke for pointing that one out). In a general case on GOV.UK it’s not going to be particularly useful as it’s also disabled for viewport width=“device-width”, but if someone’s zoomed or are using a non-standard meta viewport this might help out.