Recognise Larry’s contribution to the Cabinet Office
As Chief Mouser to the Cabinet Office (https://www.gov.uk/government/history/10-downing-street#larry-chief-mouser) Larry plays a vital part in the day-to-day running of government. Given that we’re celebrating 100 years of the Cabinet Office this month (https://www.nationalarchives.gov.uk/cabinet-office-100) it makes sense to celebrate one of its most famous postings in a small way in this style guide.