Add deploy key to let Travis push to GitHub
I've generated an SSH keypair for this repository. The private key
I encrypted using `travis encrypt-file` and committed to this repo
as `.travis/govuk_elements.enc`. The public key is `.travis/govuk_elements.pub`
which I added to the deploy keys for the repo - I gave it push access.

https://github.com/alphagov/govuk_elements/settings/keys

I've added a bit of code to the Travisfile to configure Git and SSH
and a deployment which runs a script. I renamed the script from
`jenkins.sh` (GOV.UK Jenkins used to run this script) and removed
the conditional from the script because Travis handles that for us
now.

This is better than using a `GH_TOKEN` because it's tied to a repo
rather than a user. This means that:

- if the token is compromised because of a vulnerability in Travis,
  the attacker has access to all repos that the user can access,
  not just this repo
- if the user leaves GDS, the token will stop working

I'd recommend that the `update-release-branch.sh` script is changed
to use this approach. Happy to help with that!