Fix invisible input content when changing colours
When overwriting colours via browser settings,
Firefox and IE always default to black text in form fields,
irrespective of said settings.
When the background of the form fields is transparent
and the general background is dark (e.g. black),
it results in black text on a black background,
i.e. making the text invisible and therefore completely inaccessible.

The reason why that is a problem is because
both IE and Firefox hard-code the text colour
within input fields to be the system text colour.

Unfortunately the only solution is to not set any background.
And as foreground and background colours needs to be set together,
that means The text colour cannot be `inherit` either.

Bug report for Firefox:
https://bugzilla.mozilla.org/show_bug.cgi?id=1335476