Fix broken travis deployment setup
This was broken in #408, where the setup was moved to `before_deploy`
to avoid permission issues when running PR tests in a way consistent
with other alphagov apps.

Previously that was guarded by a shell if, which was remvoved, but a
bit at the end was left and is breaking deployment builds. :(