Add comments to Travis for each deploy provider
Group these - so the first two are run on master for each commit.

The script provider is run when the version file is changed and this
generates a tag and updates the latest-release branch.

A new tag then triggers the subsequent deploy providers - to the heroku
release app and the govuk-elements-sass package to NPM.