Fix details polyfill
This fixes various issues in the details polyfill
and refactors the code a bit.

**Fix details not working with screen readers in IE**

Change the `mouseup` event to `click`
as that is the only event that JAWS and NVDA use in IE,
even though the keyboard is used.
See also http://unobfuscated.blogspot.co.uk/2013/05/event-handlers-and-screen-readers.html
This fixes #130.

**Fix enter on details' summary submitting form**

When details/summary is used within a form,
the enter key submits the form in some browsers
(e.g IE). This prevents that by
a) preventing the default action for enter and
b) changing `keyup` to `keypress` as `keyup` would be too late.

**Fix space key not opening details in Firefox**

When trying to open the details with the space key in Firefox,
they open and close again.
Not sure why that is happening, but preventing keyup fixes it.