Remove `aria-describedby` from erroring form fields
The `aria-describedby` attribute was added
before we moved the error message or hint text
for a field into its label
(or legend in case of radio/checkboxes)
in 87449f013521d09ca2b51df113f4c85c85ec36d8.

Now that it's also in the label/legend,
it gets read out multiple times to screen readers.
Therefore it is not necessary anymore and should be removed.