Create the govuk-elements-sass package
The CHANGELOG and VERSION.txt files are moved into
packages/govuk-elements-sass, as they relate to the package, not the
app.

The LICENSE.txt file is duplicated.

The README.md file is duplicated, content relating to the GOV.UK
elements app will be removed in a later commit.

The package.json file is copied to the package folder.