Make textareas wider than standard form controls
Form controls are 50% of the page width, make textures 75% of the page
width - to better fit paragraphs of text.

Add rows=“5” to increase the height of the textarea. The alternative is
to set a height - but this could be a breaking change (unless it’s a
set via a new optional class).