Improve table for numeric tabular data
Make row headers out of the first column cells
for better semantics and accessibility.

As our current default for `th` is bold
and that looked weird with the longer content which was in those cells,
this changes the content to an example which has shorter row headers.
It also changes to an example which is clearer out of context.