Move form help text into legend
Most groups of checkboxes and radio buttons within Elements
have an additional introductory help text which isn't a hint.
It's usually there to help less technical people understand
that you can select multiple checkboxes,
but can also be used for other purposes.

This moves that text into the legend to also make it accessible
to assistive technologies like screen readers.
The styling is nearly the same as before and adds a class
to the current paragraph declaration to re-use its styling.