Use h2 for error summary heading
The error summary uses an additional h1 on top of
the main h1 of the page. This is changing that to an h2.

Although it's neither inaccessible, unsemantic or wrong
to use two h1s on a page in a case like this,
it causes friction with accessibility auditors,
the good practice not to have two h1s per page,
generally people who disagree with this from a semantic point of view
and people spending time responding to such reports repeatedly.

This was user-tested with a couple of screen reader users,
as they would be the only ones affected by this change,
and the change of heading hierarchy didn't cause any problems.