Remove yarn lockfile
Set travis to use npm test rather than yarn test. We’re only using yarn
to speed up the builds for Travis, but if we can’t skip installing Yarn
during the build stages  - as it’ll install by default when a lockfile
is present in the repo - then the remove yarn lockfile and use npm
instead.