Fix summary outline shrink-wrapping in Firefox
The styling for <summary> on GOV.UK currently expects the focus ring to shrinkwrap the text. This works fine in Chrome and Safari, but fails in Firefox.

This removes the old workaround to get the arrows to show in Firefox. It then reimplements the arrows for everyone using counters. There is a potential problem that browsers that understand `counter` but not `<details>` might get another arrow in addition to the one generated by the polyfill. As luck would have it though, the list styles `disclosure-open` and `disclosure-closed` seem to have been invented for the details element, so the versions of Edge and IE that I tried this fix in don’t recognise the `::before`/`::after` `content` rules and don’t show them.

Tested in latest Firefox / Chrome / Safari / Edge and IE11. Has not been tested on mobile, other browsers or with assistive tech.