Rename 'legal text' to 'warning text'
According to #116, the 'legal text' was only called like that
to prevent people misusing the pattern and using it too often.
But the pattern is used for other than legal texts nonetheless.

This changes the name back to 'warning text' to make it closer
to what it's used for in reality and on GOV.UK.
But it also adds that it should be used sparingly.