Update GOV.UK Frontend Toolkit to 7.1.0
The details polyfill has been moved to Frontend Toolkit as part of v7.1.0. This is so that users of Elements and Frontend Toolkit can make use of it, as Elements only includes the Sass in its packages.

This was raised as part of https://github.com/alphagov/govuk_elements/issues/400