De-duplicate typography snippets
Some typography snippets exist twice,
one snippet to display the code for it
and the other to highlight something in the example above it.

This removes the example code with the highlight
as the need for the highlight is not clear,
the highlighted words look like a link,
and it is duplicate code.