Move Google Tag Manager snippet to <head>
Google Tag Manager is currently not tracking correctly as the snippet is not
placed in <head>. According to the docs, the snippet should be placed as close
to the opening <head> tag as possible, and the <noscript> snippet be placed
immediately after the opening <body> tag.

This change moves the snippet to <head> and adds a <noscript> snippet.

Source: https://developers.google.com/tag-manager/quickstart