Add txt file to reclaim account
We need to reclaim a google analytics account. To do this a file needs
to be added to the root of the site.

As we can't direct upload files to heroku and our public/ folder isnt
checked in, this is a workaround.

Once teh account is verified, the file and amended tasks will be
deleted.