Bump Travis’ node version
A recent update to Travis’ build environment [1] means that we now try to install Yarn 1.3.2 rather than 0.27.5.

Yarn introduced a check against package.json’s `engines` definition in 1.0 [2] which we are now failing, because we specify an engine of `>=8.1.4` for Heroku but install `4.4.7` on Travis:

```
[1/5] Validating package.json...

error govuk-elements@: The engine "node" is incompatible with this module. Expected version ">=8.1.4".
```

Installing the latest version of Node 8 on Travis should fix this.

[1]: https://docs.travis-ci.com/user/build-environment-updates/2017-12-12/
[2]: https://github.com/yarnpkg/yarn/pull/3675/files