Remove min, max attributes from date inputs
The current max attribute for the year is 2016, which means the examples are effectively two years ‘out of date’. Whilst services _might_ be expected to populate the max attribute dynamically, we’ve seen in user research that developers often copy and paste code examples blindly, and this is more likely to introduce bugs in the future than it is to help anyone.

The min and max attributes only really ‘apply’ when using the spinner (which we consciously disable using CSS), or when using browser form validation, which we advise against. The current experience in the browser is also odd – It’s possible to enter 2020 in the above example in desktop chrome and safari on MacOS but If you then use keyboard up/down it ‘snaps’ to 2016.

There is more discussion about this issue in #597.

Closes #597.