Remove slugignore file
In 563897a, the yarn.lock file was added to the slug ignore file, seemingly to force Heroku to install using npm rather than Yarn as 'it is more reliable'. (Files in slugignore are removed before Heroku installs dependencies [1]).

Now that we've migrated from Yarn to npm, there is no need to do this. Given that we always want to install the same packages that are in the lock file, we don't want to ignore package-lock.json, otherwise npm will just installed the latest versions permitted by package.json.

[1]: https://devcenter.heroku.com/articles/slug-compiler