Remove Travis configuration
We are unlikely to do another release from this repo, and we don't want
to have to keep a load of secrets in Travis just in case we did want to.

This commit removes all the Travis configuration, and updates the
CONTRIBUTING file to outline how to do a release without Travis.